- more changes to support unicode data
git-svn-id: svn://dev-svn.seasr.org/Components/Foundry/trunk@2731 dc7ab104-3778-4984-a5d9-166996b8e16d

+++ src-tools/org/seasr/meandre/components/tools/io/WriteFile.java
@@ -249,8 +249,9 @@ public class WriteFile extends AbstractExecutableComponent {
 
     @Override
     public void disposeCallBack(ComponentContextProperties ccp) throws Exception {
-    	if (componentContext.isFlowAborting() && file != null)
+    	if (componentContext != null && componentContext.isFlowAborting() && file != null) {
     		file.delete();
+    	}
 
     	file = null;
     }
